feature/issue#123: Move L2Pool2d into separate file
