+++ piq/functional/__init__.py
@@ -1,10 +1,12 @@
 from piq.functional.base import ifftshift, get_meshgrid, similarity_map, gradient_map
 from piq.functional.colour_conversion import rgb2lmn, rgb2xyz, xyz2lab, rgb2lab, rgb2yiq
 from piq.functional.filters import hann_filter, scharr_filter, prewitt_filter
+from piq.functional.layers import L2Pool2d
 
 
 __all__ = [
     'ifftshift', 'get_meshgrid', 'similarity_map', 'gradient_map',
     'rgb2lmn', 'rgb2xyz', 'xyz2lab', 'rgb2lab', 'rgb2yiq',
-    'hann_filter', 'scharr_filter', 'prewitt_filter'
+    'hann_filter', 'scharr_filter', 'prewitt_filter',
+    'L2Pool2d',
 ]
